requirements.txt
streamlit>=1.38
pandas>=2.2
numpy>=1.26
python-dateutil>=2.9
flake8>=7.1
